renaming App and MainApp to more meaningful terms closes #99
