+++ frontend/public/index.html
@@ -13,7 +13,7 @@
         Please enable it to continue.
       </strong>
     </noscript>
-    <div id="app"></div>
+    <div id="simple-accounting"></div>
     <!-- built files will be auto injected -->
   </body>
 </html>
