+++ frontend/src/SimpleAccounting.vue
